requirements.txt
amqp==5.3.1
ansicon==1.89.0
asgiref==3.9.1
async-timeout==5.0.1
bcrypt==4.3.0
beautifulsoup4==4.12.0
billiard==4.2.1
blessed==1.21.0
boto3==1.34.0
botocore
celery==5.3.4
cement==2.10.14
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.6
Django==4.2.7
django-cors-headers==4.3.1
django-cors-middleware==1.5.0
django-filter==23.5
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
docutils==0.19
drf-yasg==1.21.4
fabric==3.2.2
filelock==3.8.2
google==3.0.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
invoke==2.2.0
itypes==1.2.0
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.2
numpy==1.24.3
oauthlib==3.3.1
packaging==24.2
pandas==2.0.1
paramiko==4.0.0
pathspec==0.12.1
Pillow==10.1.0
pipenv==2022.12.17
platformdirs==2.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pythonnet==3.0.5
pytz==2023.3
pywin32==311
PyYAML==6.0.2
razorpay==1.3.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.7.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.9.0
semantic-version==2.10.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
soupsieve==2.4
sqlparse==0.4.3
termcolor==2.5.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
whitenoise==6.6.0
wrapt==1.17.2